organize request stubs, fix style
