+++ spec/support/request_stubs/codewars.json
