+++ spec/support/request_stubs/fake_code_wars.rb
@@ -10,6 +10,8 @@ class FakeCodeWars < Sinatra::Base
   def json_response(response_code, file_name)
     content_type :json
     status response_code
-    File.open(File.dirname(__FILE__) + "/fixtures/" + file_name, "rb").read
+    File.open(File.dirname(__FILE__) + "/" + file_name, "rb").read
   end
 end
+
+WebMock.stub_request(:any, /codewars.com/).to_rack(FakeCodeWars)
